First step in reworking tests
