+++ MangaManager/VolumeManager/VolumeManager.py
@@ -1,6 +1,7 @@
 import io
 import logging
 import os
+import pathlib
 import platform
 import re
 import tkinter as tk
@@ -24,6 +25,25 @@ logger = logging.getLogger(__name__)
 
 ScriptDir = os.path.dirname(__file__)
 
+
+def parse_fileName(filepath, volume_to_apply):
+    filename = os.path.basename(filepath)
+    regexSearch = re.findall(r"(?i)(.*)((?:Chapter|CH)(?:\.|\s)[0-9]+[.]*[0-9]*)(\.[a-z]{3})", filename)
+    if regexSearch:
+        r = regexSearch[0]
+        file_regex_finds: ChapterFileNameData = ChapterFileNameData(name=r[0], chapterinfo=r[1],
+                                                                    afterchapter=r[2], fullpath=filepath,
+                                                                    volume=volume_to_apply)
+    else:
+        # Todo: add warning no ch/chapter detected and using last int as ch identifier
+        regexSearch = re.findall(r"(?i)(.*\s)([0-9]+[.]*[0-9]*)(\.[a-z]{3}$)",
+                                 filename)  # TODO: this regex must be improved yo cover more test cases
+        if regexSearch:
+            r = regexSearch[0]
+            file_regex_finds: ChapterFileNameData = ChapterFileNameData(name=r[0], chapterinfo=r[1],
+                                                                        afterchapter=r[2], fullpath=filepath,
+                                                                        volume=volume_to_apply)
+    return file_regex_finds
 
 class App:
     def __init__(self, master: tk.Tk = None):
@@ -258,29 +278,16 @@ class App:
             else:
                 filepath = cbz_path
                 logger.debug(f"[Preview] Adding ' {filepath}' to list")
-            filename = os.path.basename(filepath)
-            regexSearch = re.findall(r"(?i)(.*)((?:Chapter|CH)(?:\.|\s)[0-9]+[.]*[0-9]*)(\.[a-z]{3})", filename)
-            if regexSearch:
-                r = regexSearch[0]
-                file_regex_finds: ChapterFileNameData = ChapterFileNameData(name=r[0], chapterinfo=r[1],
-                                                                            afterchapter=r[2], fullpath=filepath,
-                                                                            volume=volume_to_apply)
-            else:
-                # Todo: add warning no ch/chapter detected and using last int as ch identifier
-                regexSearch = re.findall(r"(?i)(.*\s)([0-9]+[.]*[0-9]*)(\.[a-z]{3}$)",
-                                         filename)  # TODO: this regex must be improved yo cover more test cases
-                if regexSearch:
-                    r = regexSearch[0]
-                    file_regex_finds: ChapterFileNameData = ChapterFileNameData(name=r[0], chapterinfo=r[1],
-                                                                                afterchapter=r[2], fullpath=filepath,
-                                                                                volume=volume_to_apply)
+
+            file_regex_finds = parse_fileName(filepath, volume_to_apply)
             new_file_path = os.path.dirname(filepath)
             if self.checkbutton_4_5_settings_val.get():
                 newFile_Name = "Filename won't be modified. Vol will be added to ComicInfo.xml"
                 file_regex_finds.complete_new_path = filepath
             else:
-                newFile_Name = f"{new_file_path}/{file_regex_finds.name} Vol.{str(volume_to_apply).zfill(2)} {file_regex_finds.chapterinfo}{file_regex_finds.afterchapter}".replace(
-                    "  ", " ")
+                newFile_Name = str(pathlib.Path(new_file_path,
+                                                f"{file_regex_finds.name} Vol.{volume_to_apply} {file_regex_finds.chapterinfo}{file_regex_finds.afterchapter}".replace(
+                                                    "  ", " ")))
                 file_regex_finds.complete_new_path = newFile_Name
 
             self._list_filestorename.append(file_regex_finds)
